+++ packages/compiler-cli/ngcc/src/dependencies/module_resolver.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 import {absoluteFrom, AbsoluteFsPath, dirname, FileSystem, isRoot, join, resolve} from '../../../src/ngtsc/file_system';
-import {PathMappings} from '../ngcc_options';
+import {PathMappings} from '../path_mappings';
 import {isRelativePath, resolveFileWithPostfixes} from '../utils';
 
 /**
